% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{api_endpoint}
\alias{api_endpoint}
\title{Create API Endpoint from Artifactory Repository}
\usage{
api_endpoint(repo)
}
\arguments{
\item{repo}{Artifactory repository; <rty.repo>}
}
\value{
<rty.api_endpoint>
}
\description{
This endpoint is particular to submitting source code to a CRAN-type repository.
}
