% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{assert_argument_class}
\alias{assert_argument_class}
\title{Asssert Argument Class}
\usage{
assert_argument_class(argument, expected_class)
}
\arguments{
\item{argument}{argument; <any>}

\item{expected_class}{expected class; <rt.one_string>}
}
\description{
Asssert Argument Class
}
