% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{available}
\alias{available}
\title{Available Packages}
\usage{
available(profile = rt::profile(), ...)
}
\arguments{
\item{profile}{Artifactory profile; <rt.profile>}

\item{...}{dot-dot-dot, passed into \code{utils::available.packages()}}
}
\value{
<matrix>
}
\description{
Available Packages
}
