% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{build_deploy}
\alias{build_deploy}
\title{Build to Artifactory}
\usage{
build_deploy(pkg = ".", profile = rt::profile(), ...)
}
\arguments{
\item{pkg}{directory path to package source code; <character>}

\item{profile}{Artifactory profile; <rt.profile>}

\item{...}{dot-dot-dot, passed into \code{devtools::build()}}
}
\description{
Build to Artifactory
}
