% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\docType{data}
\name{conditions}
\alias{conditions}
\alias{errors}
\alias{warnings}
\title{Conditions}
\format{
An object of class \code{list} of length 5.

An object of class \code{list} of length 2.
}
\usage{
errors

warnings
}
\description{
Conditions
}
\keyword{datasets}
