% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cred.R
\name{cred}
\alias{cred}
\alias{api_key}
\alias{bearer_token}
\alias{user_password}
\title{Artifactory Credentials}
\usage{
api_key(key)

bearer_token(token)

user_password(user_name, password)
}
\arguments{
\item{key}{API Key; <character>}

\item{token}{Bearer token; <character>}

\item{user_name}{User Name; <character>}

\item{password}{Password; <character>}
}
\description{
Artifactory Credentials
}
