% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{deploy}
\alias{deploy}
\title{Deploy Package to CRAN-like Repository on Artifactory}
\usage{
deploy(file, profile = rty::profile())
}
\arguments{
\item{file}{path to an R package binary file; <character>}

\item{profile}{Artifactory profile; <rty.profile>}
}
\description{
Deploy Package to CRAN-like Repository on Artifactory
}
