% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{install}
\alias{install}
\title{Install Package from CRAN-like Repository on Artifactory}
\usage{
install(name, profile = rty::profile(), version = NULL, ...)
}
\arguments{
\item{name}{Name of the package; <character>}

\item{profile}{Artifactory profile; <rty.profile>}

\item{version}{Version of the package, latest when NULL; <character, NULL>}

\item{...}{dot-dot-dot, passed into \code{utils::install.packages()}}
}
\description{
Install Package from CRAN-like Repository on Artifactory
}
