% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{one_string}
\alias{one_string}
\title{Create "one string" of Characters}
\usage{
one_string(chr)
}
\arguments{
\item{chr}{<character|!NA>}
}
\value{
<rty.one_string>
}
\description{
Create a non-NA, non-empty(""), <character> vector with length 1.
}
