% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_accessors.R
\name{option_accessors}
\alias{option_accessors}
\alias{set_repo}
\alias{get_repo}
\alias{set_cred}
\alias{get_cred}
\title{Set and Get Artifactory Profile Elements from Session Options}
\usage{
set_repo(repo, name = "default")

get_repo(name = "default")

set_cred(cred, name = "default")

get_cred(name = "default")
}
\arguments{
\item{repo}{Artifactory repository; <rt.repo>}

\item{name}{name of the element; <character>}

\item{cred}{Artifactory credentials; <rt.cred>}
}
\description{
Set and Get Artifactory Profile Elements from Session Options
}
