% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo}
\alias{repo}
\title{Artifactory Repository}
\usage{
repo(name, root)
}
\arguments{
\item{name}{Name for the repository; <character>}

\item{root}{Root URL to Arifactory instance; <character>}
}
\value{
<rt.repo>
}
\description{
Artifactory Repository
}
