% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt-package.R
\docType{package}
\name{rt-package}
\alias{rt}
\alias{rt-package}
\title{rt: Simple Helpers for Working with Artifactory}
\description{
Build and install packages to and from CRAN-like repositories on Artifactory. Set up and manage user enrivonment to keep track of repositories and credentials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iqis.github.io/rt/}
}

}
\author{
\strong{Maintainer}: Siqi Zhang \email{siqi.zhang@bayer.com}

Other contributors:
\itemize{
  \item Bayer Healthcare LLC [funder]
}

}
\keyword{internal}
