% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_rt}
\alias{with_rt}
\title{Evaluate any Expression with rt Namespace}
\usage{
with_rt(expr)
}
\arguments{
\item{expr}{expression; <any>}
}
\value{
<any>
}
\description{
rt Namespace is attached to the calling scope
}
\examples{
\dontrun{
rt::with_rt({
  install("my-package",
          profile(repo("my-repository",
                       "https://artifactory.my.domain/artifactory/"),
                  api_key("[paste API key]")))
})
}
}
