% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{Artifactory Profile}
\usage{
profile(repo = get_repo(), cred = get_cred())
}
\arguments{
\item{repo}{repository; <rty.repo>}

\item{cred}{credentials; <rty.cred>}
}
\value{
<rty.profile>
}
\description{
An Artifactory profile is the combination of Repository and Credentials.
}
