% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{Artifactory Profile}
\usage{
profile(repo = get_repo(), cred = get_cred())
}
\arguments{
\item{repo}{repository; <rt.repo>}

\item{cred}{credential; <rt.cred>}
}
\value{
<rt.profile>
}
\description{
The Artifactory profile is the combination of repository and credential.
}
